@@ -18,7 +18,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
+python = "^3.10.0"
 boto3 = "^1.26.65"
 quart = "^0.18.3"
 bento-lib = {extras = ["quart"], version = "^5.4.0"}
